# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI-Powered LMS API",
    "description": "Backend for an AI-powered Learning Management System: video ingestion and transcription, retrieval over transcripts (RAG), assessment generation, automatic answer evaluation and learning reports.\n\n**Typical flow:** `POST /videos` \u2192 poll `GET /videos/{id}` until `indexed` \u2192 `POST /videos/{id}/progress` (\u2265 threshold) \u2192 `POST /assessments` \u2192 `POST /attempts` \u2192 `GET /attempts/{id}/report`.\n\nErrors use one envelope: `{\"error\": {\"code\", \"message\", \"request_id\", \"details\"}}`.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/assessments": {
      "post": {
        "tags": [
          "assessments"
        ],
        "summary": "Generate an assessment for a watched video",
        "description": "Returns **409** unless the video is indexed and the learner's progress is at or above `COMPLETION_THRESHOLD`. Generates a mix of MCQ, true/false and short-answer questions grounded in the video's transcript. Correct answers are not included in the response.\n\n**Synchronous LLM call:** typically 20-40 s (longer while Groq rate-limits and we back off). Clients and any reverse proxy in front of the API need a read timeout of >= 120 s.",
        "operationId": "create_assessment_assessments_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssessmentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        }
      }
    },
    "/assessments/{assessment_id}": {
      "get": {
        "tags": [
          "assessments"
        ],
        "summary": "Fetch an assessment's questions",
        "operationId": "get_assessment_assessments__assessment_id__get",
        "parameters": [
          {
            "name": "assessment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Assessment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        }
      }
    },
    "/attempts": {
      "post": {
        "tags": [
          "attempts"
        ],
        "summary": "Submit answers; returns per-question evaluation and generates the report",
        "description": "MCQ and true/false are graded deterministically; short answers are graded by an LLM judge grounded in the reference answer, rubric and source transcript. Questions without a response score 0. The learning report is generated in the same transaction.\n\n**Synchronous LLM calls:** typically 5-20 s (longer while Groq rate-limits and we back off). Clients and any reverse proxy in front of the API need a read timeout of >= 120 s.",
        "operationId": "submit_attempt_attempts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttemptCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        }
      }
    },
    "/attempts/{attempt_id}": {
      "get": {
        "tags": [
          "attempts"
        ],
        "summary": "Fetch an evaluated attempt",
        "operationId": "get_attempt_attempts__attempt_id__get",
        "parameters": [
          {
            "name": "attempt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Attempt Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        }
      }
    },
    "/attempts/{attempt_id}/report": {
      "get": {
        "tags": [
          "attempts"
        ],
        "summary": "Fetch the learning report for an attempt",
        "operationId": "get_report_attempts__attempt_id__report_get",
        "parameters": [
          {
            "name": "attempt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Attempt Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnswerIn": {
        "properties": {
          "question_id": {
            "type": "string",
            "title": "Question Id"
          },
          "response": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response",
            "description": "mcq: option text, letter (A-D) or number (1-4); true_false: true/false; short_answer: free text"
          }
        },
        "type": "object",
        "required": [
          "question_id"
        ],
        "title": "AnswerIn"
      },
      "AnswerResult": {
        "properties": {
          "question_id": {
            "type": "string",
            "title": "Question Id"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          },
          "type": {
            "type": "string",
            "enum": [
              "mcq",
              "true_false",
              "short_answer"
            ],
            "title": "Type"
          },
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "response": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "is_correct": {
            "type": "boolean",
            "title": "Is Correct"
          },
          "feedback": {
            "type": "string",
            "title": "Feedback"
          },
          "improvement_areas": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Improvement Areas"
          },
          "evaluator": {
            "type": "string",
            "enum": [
              "rule",
              "llm",
              "none"
            ],
            "title": "Evaluator"
          },
          "correct_answer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correct Answer"
          },
          "reference_answer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Answer"
          }
        },
        "type": "object",
        "required": [
          "question_id",
          "position",
          "type",
          "topic",
          "prompt",
          "response",
          "score",
          "is_correct",
          "feedback",
          "improvement_areas",
          "evaluator",
          "correct_answer",
          "reference_answer"
        ],
        "title": "AnswerResult"
      },
      "AssessmentCreate": {
        "properties": {
          "video_id": {
            "type": "string",
            "title": "Video Id"
          },
          "learner_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Learner Id",
            "examples": [
              "learner-001"
            ]
          },
          "num_questions": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 20.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Num Questions"
          },
          "mix": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "propertyNames": {
                  "enum": [
                    "mcq",
                    "true_false",
                    "short_answer"
                  ]
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mix",
            "description": "Relative share per question type, e.g. {'mcq': 2, 'true_false': 1, 'short_answer': 2}"
          }
        },
        "type": "object",
        "required": [
          "video_id",
          "learner_id"
        ],
        "title": "AssessmentCreate"
      },
      "AssessmentOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "video_id": {
            "type": "string",
            "title": "Video Id"
          },
          "learner_id": {
            "type": "string",
            "title": "Learner Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/QuestionOut"
            },
            "type": "array",
            "title": "Questions"
          }
        },
        "type": "object",
        "required": [
          "id",
          "video_id",
          "learner_id",
          "created_at",
          "questions"
        ],
        "title": "AssessmentOut"
      },
      "AttemptCreate": {
        "properties": {
          "assessment_id": {
            "type": "string",
            "title": "Assessment Id"
          },
          "learner_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Learner Id"
          },
          "answers": {
            "items": {
              "$ref": "#/components/schemas/AnswerIn"
            },
            "type": "array",
            "title": "Answers"
          }
        },
        "type": "object",
        "required": [
          "assessment_id",
          "learner_id",
          "answers"
        ],
        "title": "AttemptCreate"
      },
      "AttemptOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "assessment_id": {
            "type": "string",
            "title": "Assessment Id"
          },
          "learner_id": {
            "type": "string",
            "title": "Learner Id"
          },
          "submitted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Submitted At"
          },
          "total_score": {
            "type": "number",
            "title": "Total Score"
          },
          "max_score": {
            "type": "number",
            "title": "Max Score"
          },
          "percentage": {
            "type": "number",
            "title": "Percentage"
          },
          "answers": {
            "items": {
              "$ref": "#/components/schemas/AnswerResult"
            },
            "type": "array",
            "title": "Answers"
          },
          "report_url": {
            "type": "string",
            "title": "Report Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "assessment_id",
          "learner_id",
          "submitted_at",
          "total_score",
          "max_score",
          "percentage",
          "answers",
          "report_url"
        ],
        "title": "AttemptOut"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "examples": [
              "validation_failed"
            ]
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "details": {
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorBody"
      },
      "ErrorOut": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorOut",
        "description": "The envelope every error response uses (4xx and 5xx)."
      },
...
      "QuestionOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          },
          "type": {
            "type": "string",
            "enum": [
              "mcq",
              "true_false",
              "short_answer"
            ],
            "title": "Type"
          },
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "options": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Options"
          },
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "difficulty": {
            "type": "string",
            "title": "Difficulty"
          }
        },
        "type": "object",
        "required": [
          "id",
          "position",
          "type",
          "prompt",
          "options",
          "topic",
          "difficulty"
        ],
        "title": "QuestionOut",
        "description": "A question as shown to the learner \u2014 answers and rubric are deliberately not exposed."
      },
      "ReportOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "attempt_id": {
            "type": "string",
            "title": "Attempt Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "overall": {
            "additionalProperties": true,
            "type": "object",
            "title": "Overall"
          },
          "by_type": {
            "additionalProperties": true,
            "type": "object",
            "title": "By Type"
          },
          "topic_breakdown": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Topic Breakdown"
          },
          "strengths": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Strengths"
          },
          "weaknesses": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Weaknesses"
          },
          "improvement_areas": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Improvement Areas"
          },
          "suggestions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Suggestions"
          },
          "ai_summary": {
            "type": "string",
            "title": "Ai Summary"
          },
          "summary_source": {
            "type": "string",
            "enum": [
              "llm",
              "template"
            ],
            "title": "Summary Source"
          }
        },
        "type": "object",
        "required": [
          "id",
          "attempt_id",
          "created_at",
          "overall",
          "by_type",
          "topic_breakdown",
          "strengths",
          "weaknesses",
          "improvement_areas",
          "suggestions",
          "ai_summary",
          "summary_source"
        ],
        "title": "ReportOut"
      },
...
    }
  }
}